+++ CHANGELOG.md
@@ -17,6 +17,7 @@ are solved.
 ## Unreleased
  
 - :bug: [#315](https://github.com/reasonml-editor/reasonml-idea-plugin/issues/315) ![r] insertion handler for JSX doesn't detect close tag
+- :bug: [#190](https://github.com/reasonml-editor/reasonml-idea-plugin/issues/190) show nested function definitions in structure panel
 - :nail_care: order jsx attributes, mandatory first
 
 # 0.98.3 - 2021/04/06